Development details & project completeness
We used agile methodology and divided our project into milestones: minimum functionality; low, desirable and high targets respectively, and we were able to achieve all of them within the time span we were provided. This document is the Phase 3(Final) Presentation report and includes all the functionality details of the app.